eng_hint {webex}	R Documentation

Create an engine to allow block-style hints: like “'hint button text / hint text The first line of the block is used as the button text, all other lines are rendered as markdown and included in the hidden-by-default div

Description

Create an engine to allow block-style hints: like “'hint button text / hint text The first line of the block is used as the button text, all other lines are rendered as markdown and included in the hidden-by-default div

Usage

eng_hint(options)

[Package webex version 0.0.0.9000 Index]