fitb {webex}	R Documentation

Create fill-in-the-blank question

Description

Create fill-in-the-blank question

Usage

fitb(answer, width = calculated_width, num = FALSE, tol = NULL,
  calculator = num, calculator_digits = 4, ignore_case = FALSE,
  ignore_ws = TRUE, regex = FALSE)

Arguments

answer
The correct answer

width
Width of the input box in characters. Defaults to the length of the longest answer.

num
whether the input is numeric, in which case allow for leading zeroes to be omitted

tol
the tolerance within which numeric answers will be accepts; i.e. (response - true.answer) < tol = a correct response. Implies num=TRUE

calculator
whether the input should be eval'd in the local javascript environment. Defaults to 'num“

calculator_digits
number significant digits to show

ignore_case
Whether to ignore case (capitalization)

ignore_ws
Whether to ignore white space

regex
Whether to use regex to match answers (concatenates all answers with '|' before matching)

Details

Writes html code that creates an input box widget. Call this function inline in an RMarkdown document. See the Web Exercises RMarkdown template for examples.

[Package webex version 0.0.0.9000 Index]