hide {webex}	R Documentation

Create button revealing hidden HTML content

Description

Create button revealing hidden HTML content

Usage

hide(button_text = "Solution")

Arguments

button_text
Text to appear on the button that reveals the hidden content

Details

Writes HTML to create a content that is revealed by a button press. Call this function inline in an RMarkdown document. Any content appearing after this call up to an inline call to unhide() will only be revealed when the user clicks the button. See the Web Exercises RMarkdown Template for examples.

See Also

unhide

[Package webex version 0.0.0.9000 Index]